Certificate of Participation awarded by the National Institute of Electronics and Information Technology (NIELIT)
This repository features my Certificate of Participation awarded by the National Institute of Electronics and Information Technology (NIELIT), Calicut, for the successful completion of the course "VLSI for Beginners." The 5-day online training, held from 7th to 11th February 2025, introduced foundational concepts in VLSI (Very-Large-Scale Integration) design. It was an insightful step in deepening my understanding of electronic system design and integration.